py file update

+++ run.sh
@@ -1,10 +1,12 @@
 #!/bin/sh
 
 /makeconf.sh
-mv /rs485.py /data/
+if [ ! -f /data/rs485.py ]; then
+	mv /rs485.py /data/
+fi
 
 echo "[Info] Run Wallpad Controller"
-python3 rs485.py
+python3 /data/rs485.py
 
 # for dev
-# while true; do echo "still live"; sleep 100; done
+while true; do echo "still live"; sleep 100; done
